+++ desktop/src/Umbrella.Wallet.Core/Chains/ChainId.cs
@@ -19,4 +19,5 @@ public enum ChainId
     Xrp,
     Xlm,
     Atom,
+    Near,
 }